refactor: move ReviewTableRepo out of the association package
