+++ src/main/kotlin/se/uulm/snowballr/backend/repository/ReviewTableRepo.kt
@@ -1,4 +1,4 @@
-package se.uulm.snowballr.backend.repository.association
+package se.uulm.snowballr.backend.repository
 
 import org.jetbrains.exposed.sql.ResultRow
 import org.jetbrains.exposed.sql.selectAll
@@ -6,7 +6,6 @@ import se.uulm.snowballr.backend.db.IDatabase
 import se.uulm.snowballr.backend.model.EntityType
 import se.uulm.snowballr.backend.model.SnowballRException.NotFoundException
 import se.uulm.snowballr.backend.model.dto.Review
-import se.uulm.snowballr.backend.repository.getEntityByIdOrNull
 import se.uulm.snowballr.backend.table.ReviewTable
 import se.uulm.snowballr.backend.table.toReview
 import java.util.UUID
